+++ database/migrations/2020_01_28_100000_create_loans_table.php
@@ -13,24 +13,26 @@ class CreateLoansTable extends Migration
      */
     public function up()
     {
-        Schema::create('loans', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedInteger('user_id');
-            $table->integer('amount');
-            $table->integer('terms');
-            $table->integer('outstanding_amount');
-            $table->string('currency_code');
-            $table->date('processed_at');
-            $table->string('status');
-            $table->timestamps();
-            $table->softDeletes();
-
-            $table->foreign('user_id')
-                ->references('id')
-                ->on('users')
-                ->onUpdate('cascade')
-                ->onDelete('restrict');
-        });
+        if (!Schema::hasTable("loans")) {
+            Schema::create('loans', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_id');
+                $table->integer('amount');
+                $table->integer('terms');
+                $table->integer('outstanding_amount');
+                $table->string('currency_code');
+                $table->date('processed_at');
+                $table->string('status');
+                $table->timestamps();
+                $table->softDeletes();
+    
+                $table->foreign('user_id')
+                    ->references('id')
+                    ->on('users')
+                    ->onUpdate('cascade')
+                    ->onDelete('restrict');
+            });
+        }
     }
 
     /**
